fix(worldgen): align ore generation values with Forge and re-enable dormant armor/camera code paths
In-game testing pass (16 August) confirmed most previously "shipped
without in-game test" items from PARITE_FORGE_NEOFORGE.md's pending
table, but found and fixed a real ore-generation divergence, and
re-enabled two code paths that had been left commented out.

- CNConfiguredFeatures/CNPlacedFeatures: correct ore generation values
  that didn't match Forge — URANIUM_ORE count 7→4; LEAD_ORE vein size
  10→12 and max height 112→60; THORIUM_ORE vein size 16→12, count
  16→6, max height 112→55; STRIATED_ORES_OVERWORLD rarity 1/18→1/64
  and height range -30..70→-63..16 (NITRATE_ORE was already correct,
  untouched); regenerated worldgen JSONs reflect the new values
- AntiRadiationArmorItem: uncomment the DYE_ANTI_RADIATION_ARMOR
  advancement award call in inventoryTick, now that the advancement
  itself is re-enabled (CNAdvancement)
- ClientEvents: uncomment the CameraAccessor tremor-offset camera
  move call, restoring the physical camera shake effect
- CommentEvents: register a FluidBucketWrapper item capability for
  CNFluids.URANIUM's bucket via Capabilities.FluidHandler.ITEM
- CreateNuclearDatagen: provide default lang for the "irradiated"
  translation key group
- PARITE_FORGE_NEOFORGE.md: move the uranium/graphite tooltip,
  raw_uranium_block radiation, snow-powder fan, IrradiatedWolf taming
  message, ore mining drops, enable_world_gen (both states), and the
  full nitrogen chain out of the §4 pending-verification table as
  confirmed by manual in-game testing; document the ore-generation
  bug found and fixed above with a before/after table; note that
  WolfPanicGoal and the wolf's eyeHeight fix are deliberately deferred
  to a future pass covering all of the mod's living entities rather
  than tested in isolation; and record the open-pipe uranium
  irradiation test as inconclusive, with CNOpenPipeEffectHandlers/
  RadiationEffectHandler confirmed bit-identical to Forge and
  RadiationCapability.canBeIrradiated()'s enabledItemRadiation config
  flag flagged as an unexplored lead

+++ src/main/java/net/nuclearteam/createnuclear/foundation/events/ClientEvents.java
@@ -15,7 +15,7 @@ import net.minecraft.world.entity.player.Player;
 import net.nuclearteam.createnuclear.CNClientProxy;
 import net.nuclearteam.createnuclear.CreateNuclear;
 import net.nuclearteam.createnuclear.content.equipment.armor.AntiRadiationArmorItem;
-// import net.nuclearteam.createnuclear.foundation.mixin.client.CameraAccessor;
+import net.nuclearteam.createnuclear.foundation.mixin.client.CameraAccessor;
 import net.nuclearteam.createnuclear.infrastructure.config.CNConfigs;
 
 @EventBusSubscriber(modid = CreateNuclear.MOD_ID, value = Dist.CLIENT)
@@ -72,13 +72,12 @@ public class ClientEvents {
                 double intensity = tremorAmount * Minecraft.getInstance().options.screenEffectScale().get();
 
                 // Physically offset the camera
-                /*
                 ((CameraAccessor) event.getCamera()).callMove(
-                        CNClientProxy.randomTremorOffsets[0] * 0.2F * intensity,
-                        CNClientProxy.randomTremorOffsets[1] * 0.2F * intensity,
-                        CNClientProxy.randomTremorOffsets[2] * 0.5F * intensity
+                    (float) (CNClientProxy.randomTremorOffsets[0] * 0.2F * intensity),
+                    (float) (CNClientProxy.randomTremorOffsets[1] * 0.2F * intensity),
+                    (float) (CNClientProxy.randomTremorOffsets[2] * 0.5F * intensity)
                 );
-                */
+
             }
         }
     }
